divide `zigzag.h`
