+++ include/zen/ui/nodes/board-selector/board-selector.h
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <wlr/render/wlr_renderer.h>
-#include <zigzag.h>
 
 #include "zen/screen/output.h"
+#include "zigzag/node.h"
 
 struct zn_board_selector_item_switch_button;
