+++ include/zen/ui/nodes/power-menu/power-menu.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <wlr/render/wlr_renderer.h>
-#include <zigzag.h>
+
+#include "zigzag/node.h"
 
 struct zn_power_menu_item_clock;
 struct zn_power_menu_item_logout;
