+++ include/zen/ui/nodes/vr-button.h
@@ -2,7 +2,8 @@
 
 #include <cairo.h>
 #include <wlr/render/wlr_renderer.h>
-#include <zigzag.h>
+
+#include "zigzag/node.h"
 
 struct zn_vr_menu;
