+++ include/zen/ui/nodes/vr-menu/connect-button.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <wlr/render/wlr_renderer.h>
-#include <zigzag.h>
+
+#include "zigzag/node.h"
 
 struct zn_vr_menu_headset_connect_button {
   struct zigzag_node *zigzag_node;
