+++ include/zen/ui/nodes/vr-menu/vr-menu.h
@@ -1,7 +1,8 @@
 #pragma once
 
 #include <wlr/render/wlr_renderer.h>
-#include <zigzag.h>
+
+#include "zigzag/node.h"
 
 struct zn_vr_menu_item_headset;
