+++ include/zen/ui/nodes/vr-modal/headset.h
@@ -1,9 +1,9 @@
 #pragma once
 
 #include <wlr/render/wlr_renderer.h>
-#include <zigzag.h>
 
 #include "zen/screen/output.h"
+#include "zigzag/node.h"
 
 struct zn_peer;
