+++ zen/ui/nodes/power-button.c
@@ -2,10 +2,12 @@
 
 #include <cairo.h>
 #include <zen-common.h>
-#include <zigzag.h>
 
 #include "zen/server.h"
 #include "zen/ui/layout-constants.h"
+#include "zigzag/cairo-util.h"
+#include "zigzag/layout.h"
+#include "zigzag/node.h"
 
 static int
 zn_power_button_handle_second_timer(void *data)
