+++ zen/ui/nodes/vr-menu/headset.c
@@ -2,12 +2,14 @@
 
 #include <cairo.h>
 #include <zen-common.h>
-#include <zigzag.h>
 
 #include "zen/peer.h"
 #include "zen/server.h"
 #include "zen/ui/layout-constants.h"
 #include "zen/ui/nodes/vr-menu/connect-button.h"
+#include "zigzag/cairo-util.h"
+#include "zigzag/layout.h"
+#include "zigzag/node.h"
 
 static void
 zn_vr_menu_item_headset_on_click(struct zigzag_node *node, double x, double y)
