+++ zen/ui/nodes/vr-menu/vr-menu.c
@@ -2,13 +2,15 @@
 
 #include <cairo.h>
 #include <zen-common.h>
-#include <zigzag.h>
 
 #include "zen/peer.h"
 #include "zen/server.h"
 #include "zen/ui/layout-constants.h"
 #include "zen/ui/nodes/vr-menu/connect-button.h"
 #include "zen/ui/nodes/vr-menu/headset.h"
+#include "zigzag/cairo-util.h"
+#include "zigzag/layout.h"
+#include "zigzag/node.h"
 
 static void
 zn_vr_menu_update_frame(
